Preparation for new saving
All Entities must now have a unique ID
Changed x and y to shorts since you will never be more than 32k pixels out on a map
Fixed Letter rendering

Not all Entity IDs have been implemented.

+++ src/javajesus/level/tile/Tile.java
@@ -189,7 +189,7 @@ public class Tile {
 	private static final Random random = new Random();
 
 	// unique identifier for each tile
-	private final int id;
+	private final byte id;
 
 	// whether or not an entity can walk through it
 	private final boolean solid;
@@ -217,7 +217,7 @@ public class Tile {
 	protected Tile(int id, boolean isSolid, int pixelColor, int xTile, int yTile, SpriteSheet sheet, int[] color) {
 		
 		// instance data
-		this.id = id;
+		this.id = (byte) id;
 		this.solid = isSolid;
 		this.pixelColor = pixelColor;
 		this.xTile = xTile;
@@ -236,7 +236,7 @@ public class Tile {
 	/**
 	 * @return the unique id of a tile
 	 */
-	public int getId() {
+	public byte getId() {
 		return id;
 	}
